Added basic implementation of TBlockOperator.
