+++ linalg/blockoperator.cpp
@@ -18,131 +18,6 @@
 namespace mfem
 {
 
-BlockConstOperator::BlockConstOperator(const Array<int> & offsets)
-   : Operator(offsets.Last()),
-     owns_blocks(0),
-     nRowBlocks(offsets.Size() - 1),
-     nColBlocks(offsets.Size() - 1),
-     row_offsets(offsets),
-     col_offsets(offsets),
-     op(nRowBlocks, nRowBlocks),
-     coef(nRowBlocks, nColBlocks)
-{
-   op = static_cast<const Operator *>(NULL);
-}
-
-BlockConstOperator::BlockConstOperator(const Array<int> & row_offsets_,
-                                       const Array<int> & col_offsets_)
-   : Operator(row_offsets_.Last(), col_offsets_.Last()),
-     owns_blocks(0),
-     nRowBlocks(row_offsets_.Size()-1),
-     nColBlocks(col_offsets_.Size()-1),
-     row_offsets(row_offsets_),
-     col_offsets(col_offsets_),
-     op(nRowBlocks, nColBlocks),
-     coef(nRowBlocks, nColBlocks)
-{
-   op = static_cast<const Operator *>(NULL);
-}
-
-void BlockConstOperator::SetDiagonalBlock(int iblock, const Operator *opt,
-                                          double c)
-{
-   SetBlock(iblock, iblock, opt, c);
-}
-
-void BlockConstOperator::SetBlock(int iRow, int iCol, const Operator *opt,
-                                  double c)
-{
-   if (owns_blocks && op(iRow, iCol))
-   {
-      delete op(iRow, iCol);
-   }
-   op(iRow, iCol) = opt;
-   coef(iRow, iCol) = c;
-
-   MFEM_VERIFY(row_offsets[iRow+1] - row_offsets[iRow] == opt->NumRows() &&
-               col_offsets[iCol+1] - col_offsets[iCol] == opt->NumCols(),
-               "incompatible Operator dimensions");
-}
-
-// Operator application
-void BlockConstOperator::Mult (const Vector & x, Vector & y) const
-{
-   MFEM_ASSERT(x.Size() == width, "incorrect input Vector size");
-   MFEM_ASSERT(y.Size() == height, "incorrect output Vector size");
-
-   x.Read();
-   y.Write(); y = 0.0;
-
-   xblock.Update(const_cast<Vector&>(x),col_offsets);
-   yblock.Update(y,row_offsets);
-
-   for (int iRow=0; iRow < nRowBlocks; ++iRow)
-   {
-      tmp.SetSize(row_offsets[iRow+1] - row_offsets[iRow]);
-      for (int jCol=0; jCol < nColBlocks; ++jCol)
-      {
-         if (op(iRow,jCol))
-         {
-            op(iRow,jCol)->Mult(xblock.GetBlock(jCol), tmp);
-            yblock.GetBlock(iRow).Add(coef(iRow,jCol), tmp);
-         }
-      }
-   }
-
-   for (int iRow=0; iRow < nRowBlocks; ++iRow)
-   {
-      yblock.GetBlock(iRow).SyncAliasMemory(y);
-   }
-}
-
-// Action of the transpose operator
-void BlockConstOperator::MultTranspose (const Vector & x, Vector & y) const
-{
-   MFEM_ASSERT(x.Size() == height, "incorrect input Vector size");
-   MFEM_ASSERT(y.Size() == width, "incorrect output Vector size");
-
-   x.Read();
-   y.Write(); y = 0.0;
-
-   xblock.Update(const_cast<Vector&>(x),row_offsets);
-   yblock.Update(y,col_offsets);
-
-   for (int iRow=0; iRow < nColBlocks; ++iRow)
-   {
-      tmp.SetSize(col_offsets[iRow+1] - col_offsets[iRow]);
-      for (int jCol=0; jCol < nRowBlocks; ++jCol)
-      {
-         if (op(jCol,iRow))
-         {
-            op(jCol,iRow)->MultTranspose(xblock.GetBlock(jCol), tmp);
-            yblock.GetBlock(iRow).Add(coef(jCol,iRow), tmp);
-         }
-      }
-   }
-
-   for (int iRow=0; iRow < nColBlocks; ++iRow)
-   {
-      yblock.GetBlock(iRow).SyncAliasMemory(y);
-   }
-}
-
-BlockConstOperator::~BlockConstOperator()
-{
-   if (owns_blocks)
-   {
-      for (int iRow=0; iRow < nRowBlocks; ++iRow)
-      {
-         for (int jCol=0; jCol < nColBlocks; ++jCol)
-         {
-            delete op(iRow,jCol);
-         }
-      }
-   }
-}
-
-//-----------------------------------------------------------------------
 BlockDiagonalPreconditioner::BlockDiagonalPreconditioner(
    const Array<int> & offsets_):
    Solver(offsets_.Last()),
